Add `useForeground` to Button because of DiagonalGradient
Set to true to add the ripple effect to the foreground of the view, instead of the background. This is useful if one of your child views has a background of its own, or you're e.g. displaying images, and you don't want the ripple to be covered by them.

http://facebook.github.io/react-native/docs/touchablenativefeedback.html#useforeground