fix: workaround for `isMainserverUp`
Somehow on iOS there is always `null` for `state.isInternetReachable` but we need it
to be `true` to not fallback to cached data. If there is a connection, we assume the main
server is also reachable, so we fallback to `state.isConnected` here.
With cached data on the initial app load, we cannot show anything and get stuck on the
splash screen.