refactor(settings): settings with sub screens instead of tabs
in preparation for further settings we need a more structured
settings area which we achieve with a main screen and sub screens.
this way we are not limited to tabs or have long scrolling sections.

the refactoring added the settings in sub screens that were accessible
per tabs before. all screens and sub screens are built with the one
`SettingsScreen` depending on the route param `setting`. if there
is nothing set, we render the main list, otherwise the specific
settings component.

SVA-1149